backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
cloudinary==1.28.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap-v5==1.0.5
django-bootstrap3==21.1
django-bootstrap4==21.1
django-cloudinary-storage==0.3.0
django-friendship==1.9.4
django-heroku==0.3.1
django-registration==3.1.2
django-registration-redux==2.9
django-tinymce==3.4.0
fontawesome==5.10.1.post1
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.1
lazy-object-proxy==1.3.1
Pillow==8.4.0
psycopg2==2.8.6
psycopg2-binary==2.9.2
pycodestyle==2.7.0
python-dateutil==2.7.3
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.10.11
zipp==3.6.0